Rename toJSON() as toJS() internally
BREAKING CHANGE: The `toJSON()` utility exported by 'yaml/util' is
renamed as `toJS()`.
