+++ src/ast/Collection.js
@@ -108,8 +108,7 @@ export class Collection extends Node {
     }
   }
 
-  // overridden in implementations
-  /* istanbul ignore next */
+  /* istanbul ignore next: overridden in implementations */
   toJSON() {
     return null
   }
